requirements.txt
deep-translator
pillow
